requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
astroid==3.0.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
backcall==0.2.0
bandit==1.7.5
billiard==4.1.0
black==23.9.1
celery==5.3.4
celery-redbeat==2.1.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.3
decorator==5.1.1
dill==0.3.7
diskcache==5.6.3
distlib==0.3.7
distro==1.8.0
Django==4.2.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-redis==5.4.0
django-stubs==4.2.5
django-stubs-ext==4.2.5
django-timezone-field==6.0.1
executing==2.0.0
filelock==3.12.4
flake8==6.1.0
FLAML==2.1.1
flower==2.0.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
humanize==4.8.0
identify==2.5.30
idna==3.4
ipython==8.16.1
isort==5.12.0
jedi==0.19.1
kombu==5.3.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
mypy==1.6.1
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.1
openai==1.2.2
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
pre-commit==3.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyautogen==0.1.10
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
Pygments==2.16.1
pylint==3.0.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rich==13.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.3
stevedore==5.1.0
structlog==23.2.0
tenacity==8.2.3
termcolor==2.3.0
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
types-pyOpenSSL==23.3.0.0
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-redis==4.6.0.9
types-requests==2.31.0.8
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.8
yarl==1.9.2
